ADD FlameThrower

+++ TP_TD_Train/TD_Exo_Train/Assets/TD/Scripts/Damageable.cs
@@ -5,7 +5,7 @@
 	public class Damageable : MonoBehaviour
 	{
 		[SerializeField]
-		private int _health = 1;
+		private float _health = 1;
 
 		[SerializeField]
 		private bool _destroyIfKilled = true;
@@ -16,7 +16,7 @@
 		[SerializeField]
 		private ParticleSystem _deathParticle = null;
 
-		public delegate void DamageableEvent(Damageable caller, int currentHealth, int damageTaken);
+		public delegate void DamageableEvent(Damageable caller, float currentHealth, float damageTaken);
 		private event DamageableEvent _damageTaken = null;
 
 		public event DamageableEvent DamageTaken
@@ -37,7 +37,7 @@
 			return _aimPosition.position;
 		}
 
-		public void TakeDamage(int damage)
+		public void TakeDamage(float damage)
 		{
 			_health -= damage;
